+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 2. ARM_FE - Designing and Implementing Event Creation.
 3. ARM_FE - Designing and Implementing Student Event Search by Date and Roll No.
 4. ARM_FE - Designing and Implementing Alumni Details.
+5. ARM_FE - Designing and Implementing Attendance History.